Feature/视频流分析和告警写入
待完善点：
1、截图保存没有使用OSS服务
2、模型虽然在验证集上表现较好，但是真实监控画面下， 工地工人占据画面尺寸尺寸较小、工地工人处于建筑环境之中而会被钢筋等事物遮盖，这种情况下模型表现不够好

+++ app/DB_models/camera_info_db.py
@@ -1,6 +1,8 @@
 from datetime import datetime
 import pytz
 from sqlalchemy import Column, String, DateTime, Integer
+from sqlalchemy.dialects.mssql import TINYINT
+
 from app.config.database import Base
 
 
@@ -12,7 +14,7 @@ class CameraInfoDB(Base):
     park_area = Column(String(64), nullable=False) #  摄像头所属园区区域，非空
     install_position = Column(String(64), nullable=False) #  摄像头具体安装位置，非空
     rtsp_url = Column(String(255), nullable=False) # 摄像头的RTSP地址，非空
-    analysis_mode=Column(Integer, nullable=False)# 分析模式: 0-无，1-全部，2-安全规范， 3-区域入侵， 4-火警
-    camera_status = Column(Integer, default=0)  # 摄像头状态：0-离线，1-在线
+    analysis_mode=Column(TINYINT, nullable=False)# 分析模式: 0-无，1-全部，2-安全规范， 3-区域入侵， 4-火警
+    camera_status = Column(Integer, default=0)  # 摄像头状态：0-离线，1-在线（但未开启安防检测），2-在线且安防检测中
     create_time = Column(DateTime, default=datetime.now(pytz.timezone('Asia/Shanghai'))) # 创建时间
     update_time = Column(DateTime, default=datetime.now(pytz.timezone('Asia/Shanghai')), onupdate=datetime.now(pytz.timezone('Asia/Shanghai'))) # 更新时间
